+++ src/cli/meson.build
@@ -36,6 +36,7 @@ yams_cli_sources = [
   'commands/repair_command.cpp',
   'commands/model_command.cpp',
   'commands/daemon_command.cpp',
+  'commands/daemon_status_render.cpp',
   'commands/pipeline_stage_render.cpp',
   'commands/download_command.cpp',
   'commands/watch_command.cpp',
